requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langgraph-checkpoint>=0.0.1
langgraph-checkpoint-sqlite>=0.0.1
langsmith>=0.0.77
aiosqlite>=0.19.0

# Vector store
chromadb>=0.4.22

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Optional: Media generation
replicate>=0.22.0
elevenlabs>=0.2.24

# Email delivery
resend>=0.8.0
apscheduler>=3.10.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Type checking and linting
mypy>=1.8.0
types-aiofiles>=23.2.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Production server
gunicorn>=21.2.0

# Supabase integration (optional, for storage)
supabase>=2.0.0
